# api.yaml
openapi: 3.0.1
info:
  title: Pixup Disco Microservice Endpoints
  description: Definición de los Endpoints de Disco Service para el sistema Pixup
  contact:
    name: UNAM
    url: https://www.unam.mx/
    email: rafael.martinez@cevisur.com.mx
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: artista
  description: API del Recurso Artista
paths:
  /api/artistas/{id}:
    get:
      tags:
      - artista
      summary: Obtener Artista
      operationId: artistaPorId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Artista no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
        "200":
          description: Artista encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
    put:
      tags:
      - artista
      summary: Actualizar Artista
      operationId: actualizarArtista
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artista'
        required: true
      responses:
        "404":
          description: Artista no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
        "200":
          description: Artista Actualizado Exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
    delete:
      tags:
      - artista
      summary: Eliminar Artista
      operationId: eliminarArtista
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: void
          content:
            application/json: {}
  /api/artistas:
    get:
      tags:
      - artista
      summary: Obtener todos los Artistas
      operationId: obtenerArtistas
      responses:
        "200":
          description: Listado de artistas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artista'
    post:
      tags:
      - artista
      summary: Crear Artista
      operationId: crearArtista
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artista'
        required: true
      responses:
        "201":
          description: Artista Creado Exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
components:
  schemas:
    Artista:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string